# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "API de Produtos",
    "description": "API para gerenciar produtos."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/produtos": {
      "get": {
        "tags": ["Produtos"],
        "summary": "Listar todos os produtos",
        "operationId": "listarProdutos",
        "responses": {
          "200": {
            "description": "Lista de produtos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Produtos"
                }
              }
            }
          },
          "500": {
            "description": "Um erro aconteceu"
          }
        }
      },
      "post": {
        "tags": ["Produtos"],
        "summary": "Criar um novo produto",
        "operationId": "criarProduto",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "$ref": "#/components/schemas/Produtos"
                },
                "example": {
                  "nome": "Iphone 13",
                  "descricao": "Sistema avançado de câmera dupla para fotos e vídeos incríveis",
                  "quantidade": 1,
                  "preco": 7000,
                  "desconto": 50,
                  "dataDesconto": "2023-04-25",
                  "categoria": "celular",
                  "imagemProduto": "https://br.freepik.com/fotos-vetores-gratis/iphone"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Produto criado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "message": "Criado"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação"
          },
          "500": {
            "description": "Um erro aconteceu"
          }
        }
      }
    },
    "/produtos/{id}": {
      "get": {
        "tags": ["Produtos"],
        "summary": "Obter um produto pelo ID",
        "operationId": "obterProdutoPorId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do produto",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Produto encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Produtos"
                }
              }
            }
          },
          "404": {
            "description": "Produto não encontrado"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      },

      "put": {
        "tags": ["Produtos"],
        "summary": "Atualizar um produto pelo ID",
        "operationId": "atualizarProdutoPorId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do produto a ser atualizado",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "$ref": "#/components/schemas/Produtos"
                },
                "example": {
                  "nome": "Iphone 13",
                  "descricao": "Sistema avançado de câmera dupla para fotos e vídeos incríveis",
                  "quantidade": 1,
                  "preco": 7000,
                  "desconto": 50,
                  "dataDesconto": "2023-04-25",
                  "categoria": "celular",
                  "imagemProduto": "https://br.freepik.com/fotos-vetores-gratis/iphone"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Produto editado com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "message": "Produto editado com sucesso"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Produto não encontrado"
          },
          "500": {
            "description": "Um erro aconteceu"
          }
        }
      },
      "delete": {
        "tags": ["Produtos"],
        "summary": "Deletar um produto pelo ID",
        "operationId": "deletarProdutoPorId",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do produto",
            "required": true,
            "type": "string"
          },
          {
            "name": "Produto",
            "in": "body",
            "description": "Informações do produto",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Produtos"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Produto excluído.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Produtos"
                }
              }
            }
          },
          "404": {
            "description": "Produto não encontrado"
          },
          "500": {
            "description": "Um erro aconteceu"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Produtos": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string"
          },
          "descricao": {
            "type": "string"
          },
          "quantidade": {
            "type": "integer"
          },
          "preco": {
            "type": "number"
          },
          "desconto": {
            "type": "number"
          },
          "dataDesconto": {
            "type": "string",
            "format": "date-time"
          },
          "categoria": {
            "type": "string"
          },
          "imagemProduto": {
            "type": "string"
          }
        }
      }
    }
  }
}